requirements.txt
Django==5.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-filter==24.3
python-dotenv==1.0.0